test/web-platform-tests/to-upstream/domparsing/Element-insertAdjacentHTML.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Element.prototype.insertAdjacentHTML</title>
<link rel=help href="https://w3c.github.io/DOM-Parsing/#widl-Element-insertAdjacentHTML-void-DOMString-position-DOMString-text">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {

  const el = document.createElement("p");
  el.innerHTML = "<span>test</span>";
  const foo = el.getElementsByTagName("span")[0];

  foo.insertAdjacentHTML("beforebegin", "1");
  foo.insertAdjacentHTML("afterbegin", "2");
  foo.insertAdjacentHTML("beforeend", "3");
  foo.insertAdjacentHTML("afterend", "4");

  assert_equals(el.innerHTML, "1<span>2test3</span>4");

}, "insertAdjacentHTML should insert content into the tree at the specified position");
</script>
